Working on BDD-Test for Fill Palette
